Enabled -Wextra as well.
* Removed many unused parameters in various functions.
* Removed unused looking_forward code and unused visitors.
* Moved all Looking Forward code from looking_forward_sbt to
  looking_forward.
* Made the visitors its own package, with -Wno-unused-parameter,
  to avoid unused parameter warnings.

+++ stream/sequence/src/StreamWrapper.cu
@@ -15,7 +15,7 @@ void StreamWrapper::initialize_streams(
     streams.push_back(new Stream());
   }
 
-  for (int i = 0; i < streams.size(); ++i) {
+  for (size_t i = 0; i < streams.size(); ++i) {
     streams[i]->initialize(
       number_of_events, print_memory_usage, start_event_offset, reserve_mb, i, constants, do_check);
   }
